Revert "Setup Typescript"
