+++ .gitignore
@@ -4,4 +4,3 @@ node_modules/
 package-lock.json
 src/commands/Dev/MoonCommand.js
 logs/
-dist/
