+++ package.json
@@ -4,12 +4,9 @@
   "description": "Uma bot brasileira de RPG para o Discord",
   "main": "shard.js",
   "scripts": {
-    "build": "rm -rf dist/* && babel src --extensions \".js,.ts\" --ignore *.spec.js --out-dir dist --copy-files --source-maps --verbose",
-    "start": "cross-env NODE_ENV=production node shard.js",
-    "dev": "cross-env NODE_ENV=development ts-node-dev -r tsconfig-paths/register --respawn --transpile-only --ignore-watch node_modules --no-notify src/index.ts",
+    "start": "node shard.js",
+    "dev": "nodemon shard.js",
     "lint:fix": "eslint --fix .",
-    "lint:test": "eslint .",
-    "build:test": "tsc --noEmit",
     "test": "jest"
   },
   "keywords": [],
@@ -17,7 +14,6 @@
   "dependencies": {
     "@discordjs/opus": "^0.5.0",
     "@sentry/node": "^5.24.2",
-    "babel-plugin-module-resolver": "^4.1.0",
     "bufferutil": "^4.0.3",
     "canvas": "^2.6.1",
     "dblapi.js": "^2.4.0",
@@ -42,19 +38,9 @@
   "license": "AGPL-3.0",
   "readme": "https://github.com/ySnoopyDogy/MenheraBot#readme",
   "devDependencies": {
-    "@babel/cli": "^7.14.3",
-    "@babel/core": "^7.14.3",
-    "@babel/preset-env": "^7.14.2",
-    "@babel/preset-typescript": "^7.13.0",
-    "@typescript-eslint/parser": "^4.24.0",
-    "cross-env": "^7.0.3",
     "eslint": "^7.13.0",
     "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-import-resolver-typescript": "^2.4.0",
-    "eslint-plugin-import": "^2.23.2",
-    "jest": "^26.6.3",
-    "ts-node-dev": "^1.1.6",
-    "tsconfig-paths": "^3.9.0",
-    "typescript": "^4.2.4"
+    "eslint-plugin-import": "^2.22.1",
+    "jest": "^26.6.3"
   }
 }
