+++ src/MenheraClient.js
@@ -11,11 +11,10 @@ const Constants = require('./structures/MenheraConstants');
 const RpgChecks = require('./structures/Rpgs/checks');
 const FileUtil = require('./utils/FileUtil');
 const LocaleStructure = require('./structures/LocaleStructure');
-const { ShardManager } = require('./structures/ShardManager');
 const Repositories = require('./repositories/repositories');
 
 module.exports = class MenheraClient extends Client {
-  constructor(options = {}, botSettings) {
+  constructor(options = {}) {
     super(options);
 
     this.database = Database;
@@ -26,11 +25,6 @@ module.exports = class MenheraClient extends Client {
     this.config = Config;
     this.constants = Constants;
     this.rpgChecks = RpgChecks;
-    this.botSettings = botSettings;
-
-    if (this.shard) {
-      this.shardManager = new ShardManager(this);
-    }
   }
 
   async init() {
@@ -39,8 +33,6 @@ module.exports = class MenheraClient extends Client {
     const reminder = new Reminders(this);
     reminder.loop();
     await locales.load();
-    await this.loadCommands(this.botSettings.commandsDirectory);
-    await this.loadEvents(this.botSettings.eventsDirectory);
     return true;
   }
