--- tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ES2020",
-    "lib": ["es2020"],
-    "allowJs": true,
-    "outDir": "dist",
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "baseUrl": "src",
-    "paths": {
-      "@structures/*": ["structures/*"],
-    },
-      // Ensure that .d.ts files are created by tsc, but not .js files
-    "declaration": true,
-    // Ensure that Babel can safely transpile files in the TypeScript project
-    "isolatedModules": true
-  },
-  "exclude": ["dist", "node_modules"]
-}
